refactor: Refactors is point visible
