+++ src/core/components/core-tooltip.tsx
@@ -11,6 +11,7 @@ import LiveRegion from "@cloudscape-design/components/live-region";
 import ChartSeriesDetails, { ChartSeriesDetailItem } from "../../internal/components/series-details";
 import { useSelector } from "../../internal/utils/async-store";
 import { getChartSeries } from "../../internal/utils/chart-series";
+import { isPointVisible } from "../../internal/utils/series-data";
 import { useDebouncedValue } from "../../internal/utils/use-debounced-value";
 import { ChartAPI } from "../chart-api";
 import { getFormatter } from "../formatters";
@@ -21,7 +22,6 @@ import {
   getSeriesColor,
   getSeriesId,
   getSeriesMarkerType,
-  isPointVisible,
   isXThreshold,
   matchSizeAxis,
 } from "../utils";
