+++ src/core/utils.ts
@@ -91,13 +91,6 @@ export function getBubbleSeriesSizeAxis(series: Highcharts.Series): undefined | 
   return custom?.awsui?.sizeAxis;
 }
 
-// We check point.series explicitly because Highcharts can destroy point objects, replacing the
-// contents with { destroyed: true }, violating the point's TS contract.
-// See: https://github.com/highcharts/highcharts/issues/23175.
-export function isPointVisible(point: Highcharts.Point) {
-  return point.visible && point.series && point.series.visible;
-}
-
 export function getSeriesMarkerType(series?: Highcharts.Series): ChartSeriesMarkerType {
   if (!series) {
     return "large-square";
