+++ src/internal/utils/chart-series.ts
@@ -11,15 +11,12 @@ interface InternalSeries extends Series {
   };
 }
 
+// Filters out internal series from chart, e.g., series from navigator, scrollbar, etc.
+// When calling chart.series, the result includes both the main chart series
+// and additional internal series (e.g., navigator series) when using Highstock navigator.
+// We are adding datapoints and doing other calculations and using Highstock version with a
+// navigator is causing duplicated unexpected datapoints and series entries in the chart.
 export function getChartSeries(series: InternalSeries[]) {
-  // Filters out internal series from chart, e.g., series from navigator, scrollbar, etc.
-
-  // When calling chart.series, the result includes both the main chart series
-  // and additional internal series (e.g., navigator series) when using Highstock navigator.
-
-  // We are adding datapoints and doing other calculations and using Highstock version with a
-  // navigator is causing duplicated unexpected datapoints and series entries in the chart.
-
   return series.filter((s) => !s.options.isInternal);
 }
